+++ smart-layer-network/src/main.ts
@@ -1,4 +1,5 @@
 import App from './App.svelte'
+import "@tokenscript/card-sdk/src/tokenscript.d.ts";
 
 const app = new App({
 	target: document.body,
